changed structure of neural Samer
Disclaimer: the neural Samer still doesn't work, at least for the amount of time I trained it (and on my training data).

Now, the neural Samer's neural network doesn't take up the better half of a gigabyte when stored to a file. This is because it now uses more convolution+pooling layers, so there are less values for the final dense layers to handle.